[web] Move network client to a network/ directory
* Additionally, improve enums naming.
