+++ web/src/client/network/network_manager.js
@@ -20,11 +20,11 @@
  */
 
 // @ts-check
-
-import { DBusClient } from "./dbus";
-
-import cockpit from "../lib/cockpit";
-import { int_to_text, ip4_from_text } from "../utils";
+//
+import { DBusClient } from "../dbus";
+import cockpit from "../../lib/cockpit";
+import { int_to_text, ip4_from_text } from "../../utils";
+import { NetworkEventTypes, ConnectionState } from "./index";
 
 const NM_SERVICE_NAME = "org.freedesktop.NetworkManager";
 const NM_IFACE = "org.freedesktop.NetworkManager";
@@ -32,96 +32,6 @@ const NM_SETTINGS_IFACE = "org.freedesktop.NetworkManager.Settings";
 const NM_CONNECTION_IFACE = "org.freedesktop.NetworkManager.Settings.Connection";
 const NM_ACTIVE_CONNECTION_IFACE = "org.freedesktop.NetworkManager.Connection.Active";
 const NM_IP4CONFIG_IFACE = "org.freedesktop.NetworkManager.IP4Config";
-
-const CONNECTION_ADDED = "CONNECTIONS_ADDED";
-const CONNECTION_UPDATED = "CONNECTION_UPDATED";
-const CONNECTION_REMOVED = "CONNECTION_REMOVED";
-
-/**
- * Enum for the active connection state values
- *
- * @readonly
- * @enum { number }
- * https://networkmanager.dev/docs/api/latest/nm-dbus-types.html#NMActiveConnectionState
- */
-const CONNECTION_STATE = {
-  UNKWOWN: 0,
-  ACTIVATING: 1,
-  ACTIVATED: 2,
-  DEACTIVATING: 3,
-  DEACTIVATED: 4
-};
-
-// TODO: document
-const CONNECTION_TYPES = {
-  ETHERNET: "802-3-ethernet",
-  WIFI: "802-11-wireless"
-};
-
-/**
- * @typedef {object} IPAddress
- * @property {string} address - like "129.168.1.2"
- * @property {string} prefix - like "16"
- */
-
-/**
- * @typedef {object} ActiveConnection
- * @property {string} id
- * @property {string} type
- * @property {number} state
- * @property {IPAddress[]} addresses
- * @property {string} gateway
- */
-
-/**
- * @typedef {object} Connection
- * @property {string} id
- * @property {string} name
- * @property {IPv4} ipv4
- */
-
-/**
- * @typedef {object} IPv4
- * @property {string} method
- * @property {IPAddress[]} addresses
- * @property {IPAddress[]} nameServers
- * @property {IPAddress} gateway
- */
-
-/** @typedef {(conns: ActiveConnection[]) => void} ConnectionFn */
-/** @typedef {(conns: string[]) => void} ConnectionPathsFn */
-
-/**
- * @typedef {object} Handlers
- * @property {ConnectionFn[]} connectionAdded
- * @property {ConnectionFn[]} connectionRemoved
- * @property {ConnectionFn[]} connectionUpdated
- */
-
-/**
- * @typedef {object} NetworkAdapter
- * @property {function} activeConnections
- * @property {(handler: (event: NetworkEvent) => void) => void} subscribe
- * @property {(id: string) => Promise<Connection>} getConnection
- * @property {(connection: ActiveConnection) => Promise<any>} updateConnection
- * @property {() => Promise<string>} hostname
- */
-
-/**
- * Returns given IP address in the X.X.X.X/YY format
- *
- * @param {IPAddress} addr
- * @return {string}
- */
-const formatIp = addr => `${addr.address}/${addr.prefix}`;
-
-/**
- * Network event
- *
- * @typedef {object} NetworkEvent
- * @property {string} type
- * @property {object} payload
- */
 
 /**
  * NetworkClient adapter for NetworkManager
@@ -196,7 +106,7 @@ class NetworkManagerAdapter {
     proxies.addEventListener("added", (_event, proxy) => {
       proxy.wait(() => {
         this.activeConnectionFromProxy(proxy).then(connection => {
-          handler({ type: CONNECTION_ADDED, payload: connection });
+          handler({ type: NetworkEventTypes.ACTIVE_CONNECTION_ADDED, payload: connection });
         });
       });
     });
@@ -204,13 +114,13 @@ class NetworkManagerAdapter {
     proxies.addEventListener("changed", (_event, proxy) => {
       proxy.wait(() => {
         this.activeConnectionFromProxy(proxy).then(connection => {
-          handler({ type: CONNECTION_UPDATED, payload: connection });
+          handler({ type: NetworkEventTypes.ACTIVE_CONNECTION_UPDATED, payload: connection });
         });
       });
     });
 
     proxies.addEventListener("removed", (_event, proxy) => {
-      handler({ type: CONNECTION_REMOVED, payload: proxy.path });
+      handler({ type: NetworkEventTypes.ACTIVE_CONNECTION_REMOVED, payload: proxy.path });
     });
   }
 
@@ -305,7 +215,7 @@ class NetworkManagerAdapter {
     // const settings = await this.connectionSettings(proxy.Connection);
     let addresses = [];
 
-    if (proxy.State === CONNECTION_STATE.ACTIVATED) {
+    if (proxy.State === ConnectionState.ACTIVATED) {
       const ip4Config = await this.client.proxy(NM_IP4CONFIG_IFACE, proxy.Ip4Config);
       addresses = ip4Config.AddressData.map(this.connectionIPAddress);
     }
@@ -372,123 +282,4 @@ class NetworkManagerAdapter {
   }
 }
 
-/**
- * Network client
- */
-class NetworkClient {
-  /**
-   * @param {NetworkAdapter} [adapter] - Network adapter. By default, it is set to
-o  *   NetworkManagerAdapter.
-   */
-  constructor(adapter) {
-    this.adapter = adapter || new NetworkManagerAdapter();
-    /** @type {!boolean} */
-    this.subscribed = false;
-    /** @type {Handlers} */
-    this.handlers = {
-      connectionAdded: [],
-      connectionRemoved: [],
-      connectionUpdated: []
-    };
-  }
-
-  /**
-   * Returns IP config overview - addresses, connections and hostname
-   * @return {Promise<{ addresses: IPAddress[], hostname: string, connections: Connection[]}>}
-   */
-  async config() {
-    return {
-      connections: await this.adapter.activeConnections(),
-      addresses: await this.addresses(),
-      hostname: await this.adapter.hostname()
-    };
-  }
-
-  /**
-   * Registers a callback to run when a given event happens
-   *
-   * @param {"connectionAdded" | "connectionRemoved" | "connectionUpdated"} eventType - event type
-   * @param {ConnectionFn} handler - the callback to be executed
-   * @return {function} a function to remove the callback
-   */
-  listen(eventType, handler) {
-    if (!this.subscribed) {
-      // FIXME: when/where should we unsubscribe?
-      this.subscribe();
-    }
-
-    this.handlers[eventType].push(handler);
-    return () => {
-      this.handlers[eventType].filter(h => h !== handler);
-    };
-  }
-
-  /**
-   * FIXME: improve this documentation
-   * Starts listening changes on active connections
-   *
-   * @private
-   * @return {Promise<any>} function to disable the callback
-   */
-  async subscribe() {
-    // TODO: refactor this method
-    this.subscribed = true;
-
-    this.adapter.subscribe(({ type, payload }) => {
-      switch (type) {
-        case CONNECTION_ADDED: {
-          this.handlers.connectionAdded.forEach(handler => handler(payload));
-          break;
-        }
-
-        case CONNECTION_UPDATED: {
-          this.handlers.connectionUpdated.forEach(handler => handler(payload));
-          break;
-        }
-
-        case CONNECTION_REMOVED: {
-          this.handlers.connectionRemoved.forEach(handler => handler(payload.path));
-          break;
-        }
-      }
-    });
-  }
-
-  /**
-   * Returns the active connections
-   *
-   * @returns { Promise.<Connection[]> }
-   */
-  async activeConnections() {
-    return this.adapter.activeConnections();
-  }
-
-  async getConnection(id) {
-    return this.adapter.getConnection(id);
-  }
-
-  /**
-   * Updates the connection
-   *
-   * It uses the 'path' to match the connection in the backend.
-   *
-   * @param {Connection} connection - Connection to update
-   */
-  async updateConnection(connection) {
-    return this.adapter.updateConnection(connection);
-  }
-
-  /*
-   * Returns list of IP addresses for all active NM connections
-   *
-   * @todo remove duplicates
-   * @private
-   * @return {Promise.<IPAddress[]>}
-   */
-  async addresses() {
-    const conns = await this.adapter.activeConnections();
-    return conns.flatMap(c => c.addresses);
-  }
-}
-
-export { CONNECTION_STATE, CONNECTION_TYPES, formatIp, NetworkClient, NetworkManagerAdapter };
+export { NetworkManagerAdapter };
